# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: LabTech GeoLab User Tracking API
  description: |
    Comprehensive API for user activity tracking, authentication, and administrative monitoring in the LabTech GeoLab application.
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```
    
    ## Rate Limiting
    - Standard endpoints: 100 requests per minute per IP
    - Authentication endpoints: 10 requests per minute per IP
    
    ## Error Codes
    - 400: Bad Request - Invalid input data
    - 401: Unauthorized - Missing or invalid authentication
    - 403: Forbidden - Insufficient permissions
    - 404: Not Found - Resource doesn't exist
    - 429: Too Many Requests - Rate limit exceeded
    - 500: Internal Server Error - Unexpected server error
    - 503: Service Unavailable - Database connection failed
  version: 1.0.0
  contact:
    name: LabTech GeoLab Support
    email: support@labtech-geolab.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://staging-api.labtech-geolab.com/api/v1
    description: Staging server
  - url: https://api.labtech-geolab.com/api/v1
    description: Production server
...
paths:
...
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Get a new access token using a valid refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
                  description: Valid refresh token
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login or /auth/refresh
  schemas:
...
    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: New JWT access token
        expiresIn:
          type: integer
          description: Token expiry in seconds
          example: 900
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: INVALID_INPUT
            message:
              type: string
              example: Invalid request parameters
            details:
              type: object
              additionalProperties: true
            timestamp:
              type: string
              format: date-time
            requestId:
              type: string
              format: uuid
  responses:
...
    Unauthorized:
      description: Unauthorized - Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: UNAUTHORIZED
              message: Invalid or expired token
              timestamp: '2024-01-15T10:30:00Z'
              requestId: 550e8400-e29b-41d4-a716-446655440000
...